Update compound_interest.py
return to float

+++ compound_interest.py
@@ -17,8 +17,8 @@ def compound_interest(p, t, r):
 
 
 if __name__ == "__main__":
-    p = int(input("Enter the principal amount: "))
-    t = int(input("Enter the time period: "))
-    r = int(input("Enter the rate of interest: "))
+    p = float(input("Enter the principal amount: "))
+    t = float(input("Enter the time period: "))
+    r = float(input("Enter the rate of interest: "))
 
     print("The compound interest is {:.2f}".format(compound_interest(p, t, r)))
